start work on main menu casual stats
This commit should probably be undone when work is continued.

The goal is to put NameLabel, DescLine1, DescLine2, as well as probably the XP bar, to the right of the rank icon.

I should probably read or watch a hud guide to understand what is going on and how to do what I want.

The changes regarding the stats being shown are based and should be kept. I just need to check whether all the things that have had their visibility changed to 1 actually need to.